+++ AUTOFIX_backend/src/main/java/com/autofix/AUTOFIX_backend/entities/RecargoKilometrajeEntity.java
@@ -18,8 +18,8 @@ public class RecargoKilometrajeEntity {
     @Column(unique = true, nullable = false)
     private Long id;
 
-    private double cotaInicial;
-    private double cotaFinal;
+    private Long cotaInicial;
+    private Long cotaFinal;
     private int sedan;
     private int hatchback;
     private int suv;
